+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - created the ability to send statistical data to Omniva PowerBi
 - added that after the formatting function, the type of the measurement values is changed to string to avoid the problem when some servers provide a value with many numbers after the decimal point when converting to json
 - added COD amount value to be converted to string to avoid value with many decimal numbers issue on some servers
+- CallCourierOmxRequest (and CallCourier for backwards compatibility) now accepts timezone for better calculation of pickup datetime, if not set uses server timezone
 
 ## [1.1.0] - Improvements
 - adapted to work with the Omniva OMX server